closes 7 by moving streamers to api client making them available in the nuget pkg as well.
