+++ TastyTrade.Client.Examples/Program.cs
@@ -1,6 +1,10 @@
-﻿using System.Threading;
+﻿using System.IO;
+using System.Text.Json;
+using System.Threading;
 using System.Threading.Tasks;
 using DxFeed.Graal.Net;
+using TastyTrade.Client.Model.Request;
+using TastyTrade.Client.Repository;
 
 namespace TastyTrade.Client.Examples;
 
@@ -12,8 +16,9 @@ static class Program
     {
         SystemProperty.SetProperty("dxfeed.experimental.dxlink.enable", "true");
         SystemProperty.SetProperty("scheme", "ext:opt:sysprops,resource:dxlink.xml");
-        
-        await OptionChainStreamer.Run();
+
+        var credentials = JsonSerializer.Deserialize<AuthorizationCredentials>(await File.ReadAllTextAsync(CredsPath));
+        await OptionChainStreamer.Run(credentials);
         await Task.Delay(Timeout.Infinite);
     }
 }
