+++ TastyTrade.Client/Repository/RepositoryConstants.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TastyTrade.Client.Repository
+{
+    public class RepositoryConstants
+    {
+        /// <summary>
+        /// TODO: remove this and load credentials via injection OR calling parameter into repository classes
+        /// </summary>
+        public const string DefaultCredentialsPath = "./credentials.json";
+    }
+}
