+++ TastyTrade.Client/Repository/Streaming/AccountStreamer.cs
@@ -15,6 +15,7 @@ using TastyTrade.Client.Model.Request.Streaming;
 using TastyTrade.Client.Model.Response.Streaming;
 using TastyTrade.Client.Model.Helper;
 using TastyTrade.Client.Utils;
+using TastyTrade.Client.Repository;
 
 namespace TastyTrade.Client.Examples
 {
@@ -79,10 +80,17 @@ namespace TastyTrade.Client.Examples
         }
         public static async Task Run()
         {
+            var credentials = JsonSerializer.Deserialize<AuthorizationCredentials>(await File.ReadAllTextAsync(RepositoryConstants.DefaultCredentialsPath));
+            await Run(credentials);
+        }
+
+        public static async Task Run(AuthorizationCredentials credentials)
+        {
+
             _accountUpdates = new AccountDataUpdates();
 
             SetupSerializationAttributeStringMaps();
-            var credentials = JsonSerializer.Deserialize<AuthorizationCredentials>(await File.ReadAllTextAsync(Program.CredsPath));
+            
             var tastyTradeClient = new TastyTradeClient();
             await tastyTradeClient.Authenticate(credentials);
